+++ src/containers/Generation/context/GenerationContext.test.jsx
@@ -37,6 +37,9 @@ describe('Generation Context', () => {
   const mockAssignmentRows = JSON.parse(
     '[{"contract": "ES0031405524755001RN0F - 0010777","contract_address": "Major, 22, 3º 08970 (Sant Joan Despí)","priority": 0,"contract_last_invoiced": "2023-01-08","annual_use_kwh": "7105.0"},{"contract": "ES0031405524910014WM0F - 0013117","contract_address": ". Jacint Verdaguer, 42, 3er 1a 8970 (Sant Joan Despí)","priority": 1,"contract_last_invoiced": "2023-01-04","annual_use_kwh": "115.0"}]'
   )
+  const mockOutsideAssignmentRows = JSON.parse(
+    '[{"contract": "ES0031405524755001RN0F - 0010777","contract_address": "Major, 22, 3º 08970 (Sant Joan Despí)","priority": 0,"contract_last_invoiced": "2023-01-08","annual_use_kwh": "7105.0"},{"contract": "ES0031405524910014WM0F - 0013117","contract_address": ". Jacint Verdaguer, 42, 3er 1a 8970 (Sant Joan Despí)","priority": 1,"contract_last_invoiced": "2023-01-04","annual_use_kwh": "115.0"}]'
+  )
   const mockInvestmentRows = JSON.parse(
     '[{"name": "GKWH05844","nominal_amount": "2000.0","nshares": 20,"purchase_date": "2020-01-01","first_effective_date": "2021-01-01","amortized_amount": "160.0","last_effective_date": "2045-01-01"}]'
   )
@@ -45,7 +48,8 @@ describe('Generation Context', () => {
     const dom = render(
       <GenerationContextProvider
         assignmentsJSON={mockAssignmentRows}
-        investmentsJSON={mockInvestmentRows}>
+        investmentsJSON={mockInvestmentRows}
+        outsideAssignmentsJSON={mockOutsideAssignmentRows}>
           <ContextConsumer />
         </GenerationContextProvider>
     )
@@ -59,7 +63,8 @@ describe('Generation Context', () => {
     const dom = render(
       <GenerationContextProvider
         assignmentsJSON={mockAssignmentRows}
-        investmentsJSON={mockInvestmentRows}>
+        investmentsJSON={mockInvestmentRows}
+        outsideAssignmentsJSON={mockOutsideAssignmentRows}>
           <ContextConsumer />
         </GenerationContextProvider>
     )
@@ -74,7 +79,8 @@ describe('Generation Context', () => {
     const dom = render(
       <GenerationContextProvider
         assignmentsJSON={mockAssignmentRows}
-        investmentsJSON={mockInvestmentRows}>
+        investmentsJSON={mockInvestmentRows}
+        outsideAssignmentsJSON={mockOutsideAssignmentRows}>
           <ContextConsumer />
         </GenerationContextProvider>
     )
@@ -95,7 +101,8 @@ describe('Generation Context', () => {
     render(
       <GenerationContextProvider
         assignmentsJSON={mockAssignmentRows}
-        investmentsJSON={mockInvestmentRows}>
+        investmentsJSON={mockInvestmentRows}
+        outsideAssignmentsJSON={mockOutsideAssignmentRows}>
           <ContextConsumer />
         </GenerationContextProvider>
     )